+++ NEWS.md
@@ -1,10 +1,14 @@
-# betafunctions v. 1.8.2
+# betafunctions v. 1.9.0
 
 - Fixed a bug with fitting the two-parameter beta-distribution model in the `HB.CA()`, `HB.CA.MC()`, and `HB.ROC()` functions.
 
-- The `AMS()` and `BMS()` functions have had some arguments removed. Now only allows for specifying mean, variance, and location parameters.
+- The `AMS()` and `BMS()` functions have had some arguments removed. Now only allows for specifying mean, variance (or standard deviation), and location parameters.
+
+- The `LABMSU()` and `UABMSL()` functions have had arguments removed due to arguments being removed from `AMS()` and `BMS()` (the `skewness` and `kurtosis` arguments in particular.
 
 - The `Beta.tp.fit()` function is significantly simplified and some arguments are removed (the `alpha` and `beta` arguments, specifically).
+
+- The `d/p/q/rBetaMS()` functions now includes arguments for specifying lower- and upper-bound location parameters. This allows the mean to be specified outside the [0, 1] interval (provided the specified mean is located within the lower and upper bounds).
 
 ---